+++ src/table/Table.tsx
@@ -73,6 +73,14 @@ type TableProps = {
    * allow to specify extra props for the search input
    */
   searchProps?: any;
+  /**
+   * allow to specify a custom header labels
+   */
+  customHeaderLabels?: { label: string; data: string }[];
+  /**
+   * allow to pass extra properties to each row
+   */
+  trProps?: any[];
 };
 
 const keys = (dataSource: any[], columnsToOmit?: string[]): string[] =>
@@ -96,6 +104,8 @@ export const Table = ({
   withOrderBy = false,
   withSearch = false,
   searchProps,
+  customHeaderLabels,
+  trProps,
 }: TableProps) => {
   const { items, requestSort, sortConfig } = useSortableData(dataSource);
   const [selectedHeader, setSelectedHeader] = React.useState('');
@@ -108,8 +118,16 @@ export const Table = ({
   });
 
   const handleClick = (value: string) => {
-    requestSort(value);
-    setSelectedHeader(value);
+    if (customHeaderLabels) {
+      const customHeaderEl = customHeaderLabels.find(c => c.label === value);
+      //@ts-ignore
+      requestSort(customHeaderEl.data);
+      //@ts-ignore
+      setSelectedHeader(customHeaderEl.data);
+    } else {
+      requestSort(value);
+      setSelectedHeader(value);
+    }
   };
 
   const getClassNamesFor = (name: string) => {
@@ -138,7 +156,11 @@ export const Table = ({
           theadClassName={theadClassName}
           trClassName={trClassName}
           thClassName={thClassName}
-          values={keys(items, columnsToOmit)}
+          values={
+            customHeaderLabels
+              ? Array.from(customHeaderLabels, c => c.label)
+              : keys(items, columnsToOmit)
+          }
           keySorted={getClassNamesFor(selectedHeader)}
           sortAscIcon={sortAscIcon}
           sortDescIcon={sortDescIcon}
@@ -153,6 +175,7 @@ export const Table = ({
           tbodyClassName={tbodyClassName}
           trClassName={trClassName}
           tdClassName={tdClassName}
+          trProps={trProps}
         />
       </table>
     </>