+++ CHANGELOG.md
@@ -1,20 +1,38 @@
+
+## 0.3.0 (date tbd)
+
+<a name="0.3.0"></a>
+
+- [Accessibility issues fixed against all current components](https://github.com/Capgemini/dcx-react-library/issues/200)
+- [toolTip](https://github.com/Capgemini/dcx-react-library/issues/199)
+- [table with custom rows](https://github.com/Capgemini/dcx-react-library/issues/221)
+
 ## 0.2.4 (06/10/21)
 
 <a name="0.2.4"></a>
 
 - [Add anchor to each tab](https://github.com/Capgemini/dcx-react-library/issues/229)
-## 0.2.3 (07/07/21)
+
+## 0.2.3 (08/07/2021)
 
 <a name="0.2.3"></a>
 
-- [Table should display also if there's no data](https://github.com/Capgemini/dcx-react-library/issues/218)
-## 0.2.2 (07/07/21)
+- [table with custom header](https://github.com/Capgemini/dcx-react-library/issues/218)
+
+## 0.2.2 (07/07/2021)
 
 <a name="0.2.2"></a>
 
 - [select with value](https://github.com/Capgemini/dcx-react-library/issues/213)
 - [select with nullOption](https://github.com/Capgemini/dcx-react-library/issues/210)
-## 0.2.0 (02/07/21)
+
+## 0.2.1 (06/07/2021)
+
+<a name="0.2.1"></a>
+
+- [select with nullOption](https://github.com/Capgemini/dcx-react-library/issues/210)
+
+## 0.2.0 (02/07/2021)
 
 <a name="0.2.0"></a>
 
@@ -27,12 +45,12 @@
 
 <a name="0.1.2"></a>
 
-## 0.1.2
+## 0.1.2 (02/06/21)
 
 - [ForgJs vulnerability](https://github.com/Capgemini/dcx-react-library/issues/201)
 
 <a name="0.1.1"></a>
 
-# 0.1.1 (02/05/21)
+## 0.1.1 (02/05/21)
 
 - [Button ClassName and optional Label](https://github.com/Capgemini/dcx-react-library/issues/186)